+++ modext/windup/core.py
@@ -1,6 +1,6 @@
 
 from modcore.log import LogSupport
-from modcore.fiber import FiberLoop, Fiber
+from modext.fiber import FiberLoop, Fiber
 
 from modext.http.webserv import WebServer, COOKIE_HEADER, SET_COOKIE_HEADER
 from modext.http.http_func import BadRequestException